cli_backend/requirements.txt
pandas
requests
sqlalchemy
psycopg2-binary
simple-salesforce
pytest
pymongo
kafka-python
